Add event return listeners for Twitter warning,limit and discnct
considering the amount of information you will be processing it
is more imperative for you to take into account and work on being
able to keep track of what twitters feedback is going to be. Thus
created a series of event listeners for message feedback from
twitter in the case.