Add function for when a tweet is received and process to client
setup function for when a tweet is received and how to clean it
up and validate it. Converted the text to lowercase though will
have to figure out what to do on clientside with that
and then checked if any of the topics is contained in the
tweet text and then sends a cleaned up version of the tweet.